+++ src_hw/layers/softmax_layer.cpp
@@ -12,12 +12,12 @@ SoftmaxLayer::SoftmaxLayer(std::string name, std::string bottom, std::string top
 	Logger::GetLogger()->LogMessage("\tSoftmax layer '%s' constructed with bottom = '%s' and top = '%s'", name.c_str(), bottom.c_str(), top.c_str());
 }
 
-void SoftmaxLayer::Reshape(const Blob<float>* bottom, Blob<float>* top)
+void SoftmaxLayer::Reshape(const Blob<int>* bottom, Blob<int>* top)
 {
 	top->ReshapeLike(*bottom);
 }
 
-void SoftmaxLayer::Forward(const Blob<float>* bottom, Blob<float>* top)
+void SoftmaxLayer::Forward(const Blob<int>* bottom, Blob<int>* top)
 {
 	Logger::GetLogger()->LogMessage("\t%s layer performing forward computation", name_.c_str());
 
@@ -36,13 +36,13 @@ bool SoftmaxTest() // TODO test proper version
 	int count = num * channels * height * width;
 
 	SoftmaxLayer prob1("prob_test", "test_in", "test_out"); // initialise dropout layer
-	Blob<float> bottomBlob(num, channels, height, width);
-	Blob<float> topBlob;
+	Blob<int> bottomBlob(num, channels, height, width);
+	Blob<int> topBlob;
 
 	prob1.SetUp(&bottomBlob, &topBlob);
 
 	// set input data
-	float *dataIn = new float[count];
+	int *dataIn = new int[count];
 	for(int dataIndex = 0; dataIndex < count; dataIndex++)
 	{
 		dataIn[dataIndex] = dataIndex * -1^dataIndex;
@@ -52,8 +52,8 @@ bool SoftmaxTest() // TODO test proper version
 	prob1.Forward(&bottomBlob, &topBlob); // perform forward computation
 
 	// get results
-	const float* bottomData = bottomBlob.getConstData();
-	const float* topData = topBlob.getConstData();
+	const int* bottomData = bottomBlob.getConstData();
+	const int* topData = topBlob.getConstData();
 
 	// check results
 	bool testPassed = true;
